refactor: extract per-tool packages + shared whisper worker
@mtosity/tool-speech-to-text, @mtosity/tool-img-grid, @mtosity/tool-instagram
own their UI, libs (yt-dlp, instagram-url) and API handlers; web app keeps
thin route files (segment configs stay literal for static analysis).
Shared Whisper worker lives in @mtosity/whisper with per-tool worker entries.
