+++ apps/web/package.json
@@ -17,6 +17,9 @@
     "@huggingface/transformers": "^4.2.0",
     "@mtosity/design-system": "workspace:*",
     "@mtosity/lib": "workspace:*",
+    "@mtosity/tool-img-grid": "workspace:*",
+    "@mtosity/tool-instagram": "workspace:*",
+    "@mtosity/tool-speech-to-text": "workspace:*",
     "@lottiefiles/react-lottie-player": "^3.5.4",
     "@neondatabase/serverless": "^1.1.0",
     "@tiptap/extension-image": "^3.26.0",
