+++ packages/tool-img-grid/src/components/LayoutPreview.tsx
