+++ packages/tool-img-grid/src/index.ts
@@ -1,0 +1,1 @@
+export { default as ImgGridTool } from "./ImgGridTool";
