+++ packages/tool-instagram/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@mtosity/tool-instagram",
+  "version": "0.0.0",
+  "private": true,
+  "exports": {
+    ".": "./src/index.ts",
+    "./api/info": "./src/api/info.ts",
+    "./api/audio": "./src/api/audio.ts",
+    "./api/download": "./src/api/download.ts",
+    "./api/thumb": "./src/api/thumb.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@mtosity/design-system": "workspace:*",
+    "@mtosity/lib": "workspace:*",
+    "@mtosity/whisper": "workspace:*",
+    "framer-motion": "catalog:"
+  },
+  "peerDependencies": {
+    "next": ">=16",
+    "react": "^18.0.0 || ^19.0.0",
+    "react-dom": "^18.0.0 || ^19.0.0"
+  },
+  "devDependencies": {
+    "@mtosity/typescript-config": "workspace:*",
+    "@types/node": "^24.0.0",
+    "@types/react": "catalog:",
+    "@types/react-dom": "catalog:",
+    "next": "catalog:",
+    "react": "catalog:",
+    "react-dom": "catalog:",
+    "typescript": "catalog:"
+  }
+}
