+++ packages/tool-instagram/src/index.ts
@@ -1,0 +1,1 @@
+export { default as InstagramTool } from "./InstagramTool";
