+++ packages/tool-instagram/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@mtosity/typescript-config/react-library.json",
+  "compilerOptions": {
+    "lib": ["dom", "dom.iterable", "esnext", "webworker"]
+  },
+  "include": ["src"],
+  "exclude": ["node_modules"]
+}
