+++ packages/tool-speech-to-text/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "@mtosity/tool-speech-to-text",
+  "version": "0.0.0",
+  "private": true,
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@mtosity/design-system": "workspace:*",
+    "@mtosity/whisper": "workspace:*",
+    "framer-motion": "catalog:"
+  },
+  "peerDependencies": {
+    "next": ">=16",
+    "react": "^18.0.0 || ^19.0.0",
+    "react-dom": "^18.0.0 || ^19.0.0"
+  },
+  "devDependencies": {
+    "@mtosity/typescript-config": "workspace:*",
+    "@types/react": "catalog:",
+    "@types/react-dom": "catalog:",
+    "next": "catalog:",
+    "react": "catalog:",
+    "react-dom": "catalog:",
+    "typescript": "catalog:"
+  }
+}
