+++ packages/tool-speech-to-text/src/index.ts
@@ -1,0 +1,1 @@
+export { default as SpeechToTextTool } from "./SpeechToTextTool";
