+++ packages/tool-speech-to-text/src/whisper-worker.ts
@@ -1,0 +1,4 @@
+// Worker entry point. Webpack requires `new Worker(new URL(...))` to use a
+// path relative to the importing module, so each tool ships this thin entry
+// that pulls in the shared whisper worker implementation.
+import "@mtosity/whisper/worker";
