+++ packages/whisper/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@mtosity/whisper",
+  "version": "0.0.0",
+  "private": true,
+  "exports": {
+    "./worker": "./src/worker.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@huggingface/transformers": "^4.2.0"
+  },
+  "devDependencies": {
+    "@mtosity/typescript-config": "workspace:*",
+    "typescript": "catalog:"
+  }
+}
