+++ packages/whisper/src/worker.ts
