+++ packages/whisper/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@mtosity/typescript-config/base.json",
+  "compilerOptions": {
+    "lib": ["esnext", "webworker"]
+  },
+  "include": ["src"],
+  "exclude": ["node_modules"]
+}
